+++ hw12/go.mod
@@ -1,4 +1,4 @@
-module github.com/ayupov-ayaz/otus-wh-test/hw12_13_14_15_calendar
+module github.com/ayupov-ayaz/otus-wh-test/hw12
 
 go 1.16
