+++ hw12/internal/storage/entity/event.go
