+++ hw12/internal/storage/event/memory/storage.go
@@ -2,10 +2,11 @@ package memory
 
 import (
 	"context"
-	"github.com/ayupov-ayaz/otus-wh-test/hw12_13_14_15_calendar/internal/storage/entity"
 	"sync"
 
-	"github.com/ayupov-ayaz/otus-wh-test/hw12_13_14_15_calendar/internal/storage"
+	"github.com/ayupov-ayaz/otus-wh-test/hw12/internal/storage/entity"
+
+	"github.com/ayupov-ayaz/otus-wh-test/hw12/internal/storage"
 )
 
 type Storage struct {
