+++ hw12/internal/storage/event/mysql/storage.go
@@ -5,10 +5,11 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"github.com/ayupov-ayaz/otus-wh-test/hw12_13_14_15_calendar/internal/storage"
-	"github.com/ayupov-ayaz/otus-wh-test/hw12_13_14_15_calendar/internal/storage/entity"
-	"github.com/jmoiron/sqlx"
 	"time"
+
+	"github.com/ayupov-ayaz/otus-wh-test/hw12/internal/storage"
+	"github.com/ayupov-ayaz/otus-wh-test/hw12/internal/storage/entity"
+	"github.com/jmoiron/sqlx"
 )
 
 const (
