+++ hw12/internal/storage/event/mysql/storage_test.go
@@ -5,15 +5,16 @@ package mysql
 
 import (
 	"context"
-	store "github.com/ayupov-ayaz/otus-wh-test/hw12_13_14_15_calendar/internal/storage"
-	"github.com/ayupov-ayaz/otus-wh-test/hw12_13_14_15_calendar/internal/storage/entity"
-	"github.com/ayupov-ayaz/otus-wh-test/hw12_13_14_15_calendar/internal/storage/test"
-	"github.com/jmoiron/sqlx"
-	"github.com/stretchr/testify/require"
 	"os"
 	"strconv"
 	"testing"
 	"time"
+
+	store "github.com/ayupov-ayaz/otus-wh-test/hw12/internal/storage"
+	"github.com/ayupov-ayaz/otus-wh-test/hw12/internal/storage/entity"
+	"github.com/ayupov-ayaz/otus-wh-test/hw12/internal/storage/test"
+	"github.com/jmoiron/sqlx"
+	"github.com/stretchr/testify/require"
 )
 
 var (
