+++ hw12/internal/storage/storage.go
@@ -3,7 +3,8 @@ package storage
 import (
 	"context"
 	"errors"
-	"github.com/ayupov-ayaz/otus-wh-test/hw12_13_14_15_calendar/internal/storage/entity"
+
+	"github.com/ayupov-ayaz/otus-wh-test/hw12/internal/storage/entity"
 )
 
 var ErrEventNotFound = errors.New("event not found")
