+++ hw12/internal/storage/user/mysql/storage.go
@@ -2,7 +2,8 @@ package mysql
 
 import (
 	"context"
-	"github.com/ayupov-ayaz/otus-wh-test/hw12_13_14_15_calendar/internal/storage/entity"
+
+	"github.com/ayupov-ayaz/otus-wh-test/hw12/internal/storage/entity"
 	"github.com/jmoiron/sqlx"
 )
